Re-worked the nurseries ui
The nursery UI now matches the new farm UI for propogated plants and
orders.

Closes #15.
